Inline lint
